Added support for rotation to entities

+++ src/main/java/omnivoxel/client/game/graphics/opengl/mesh/vertex/TextureVertex.java
@@ -1,4 +1,4 @@
 package omnivoxel.client.game.graphics.opengl.mesh.vertex;
 
-public record TextureVertex(int tx, int ty) {
+public record TextureVertex(float tx, float ty) {
 }
